pimd: Always have a valid pim_interface object set in ifp->info
All other FRR daemons follow the standard that the ifp->info object is
created when the interface object itself is created, using the if_add
hook. PIM was not doing this and only creating the pim_interface object
when pim or gm was enabled.
This commit changes that behavior so that pim_interface is always
created when the interface object is created and when pim/gm is
enabled/disabled, it will set the values in pim_interface apporiately.
This way you can count on the pim_interface existing if the interface exists.

Signed-off-by: Nathan Bahr <nbahr@atcorp.com>